requirements.txt
aiomysql==0.1.1
aioredis==2.0.1
anyio==3.6.1
APScheduler==3.9.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
click==8.1.3
Cython==0.29.32
dateutils==0.6.12
DBUtils==3.0.2
Deprecated==1.2.13
fastapi==0.85.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
loguru==0.6.0
packaging==21.3
pydantic==1.10.2
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
rfc3986==1.5.0
selectolax==0.3.11
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
uvicorn==0.18.3
wrapt==1.14.1
